minimize changes skeleton II
